+++ resources/js/layout/Spinner/InfiniteLoader.scss
@@ -1,0 +1,4 @@
+.__infinite-loader {
+  margin: 20px;
+  text-align: center;
+}
