+++ resources/js/routes/connections/Connections.scss
@@ -169,11 +169,6 @@
       margin-bottom: 8px;
     }
   }
-
-  .infinite-loader {
-    margin: 20px;
-    text-align: center;
-  }
 }
 
 .__connection-type-dropdown {
